Backend/requirements.txt
fastapi
motor
chess
openai
python-dotenv
prettytable
Pillow